laser wall 충돌 구현 중 1

+++ src/pack_001/Obstacle.java
@@ -29,14 +29,14 @@ interface Wall extends Obstacle {
 abstract class Colider {
 	float x, y; // 오브젝트 화면 위치
 	int cx = 0, cy = 0; // 충돌 박스의 오브젝트에 대한 상대 위치
-	int cwidth = 0, cheight = 0; // 충돌 박스 크기
+	float cwidth = 0, cheight = 0; // 충돌 박스 크기
 	protected boolean trigger = false; // 충돌 가능 여부
 	
 	public Colider(){
 		
 	}
 	
-	public void setBox(int x, int y, int width, int height){
+	public void setBox(int x, int y, float width, float height){
 		this.cx = x;
 		this.cy = y;
 		this.cwidth = width;
